adds post sign up view leader board button
